fix: 자동차 경주 리팩토링
